Daniela ab

+++ resorte.cpp
@@ -2,7 +2,7 @@
 
 
 
-double resorte::getVelocidad() const
+float resorte::getVelocidad() const
 {
     return velocidad;
 }
@@ -97,3 +97,4 @@ void resorte::sprint()
 {
 
 }
+
